requirements.txt
scapy
numpy
scikit-learn
joblib
flask
flask-restful
flask-jwt-extended
flask-socketio
sqlalchemy
requests
gunicorn
eventlet
flask-cors
python-dotenv
pandas
psutil
